Stop setting a `DashDocSetFallbackURL` in `Info.plist`
We can no longer use this approach to support the "Open Online Page" feature, because the directory structure on the ocaml.org website has diverged from the directory structure of the .tar.gz manual.

e.g. the path `htmlman/libref/Option.html` does not appear in the URL `https://v2.ocaml.org/releases/4.14/api/Option.html`

---

Funnily enough, Dash's default behaviour for the "Open Online Page" feature (when we don't set a `DashDocSetFallbackURL`) actually works.

It goes through a redirect on the Kapeli server (which presumably doesn't know about our Unofficial OCaml Docset, but does know about the `ocaml` "platform", which we are associated with because we set `DocSetPlatformFamily` in `Info.plist`):

	https://kapeli.com/dash_share?docset_file=ocaml-unofficial&docset_name=OCaml%20(Unofficial)&path=htmlman/libref/Option.html&platform=ocaml

...that then goes to the following URL on the Inria server (which itself is another redirect):

	https://caml.inria.fr/pub/docs/manual-ocaml/libref/Option.html

...and finally ends up on:

	https://ocaml.org/api/Option.html

...which is not versioned (i.e. there's no `4.14` in that URL, so it's presumably whatever the latest version is) but at least it works, so is better than nothing.