Deno Deploy: Delete Makefile target to run locally; Update todo
There's no point testing locally - it's too prone to false positive. Because often something that works in Deno-CLI ultimately doesn't work in Deno-Deploy, either because the feature/API simply won't be supported in Deploy at all, or because development of the Deploy runtime is lagging behind in some regard. So just test using Preview Deployments instead.